Refactor Settings screen and update navigation
- Moved the Settings screen to a new file for better organization.
- Updated the import path for the Settings screen in BottomTabs.
- Added a new Settings component with logout functionality and pending forms handling.
- Enhanced the Login screen by adding a missing import statement.
