+++ src/app/screens/settings/Settings.tsx
@@ -4,16 +4,16 @@ import { SafeAreaView } from 'react-native-safe-area-context';
 import { TouchableOpacity } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import { useTranslation } from 'react-i18next';
-import LanguageControl from '../components/LanguageControl';
+import LanguageControl from '../../components/LanguageControl';
 import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
-import { getQueue } from '../pendingQueue';
-import { SubmissionItem } from '../../types';
+import { getQueue } from '../../pendingQueue';
+import { SubmissionItem } from '../../../types';
 import { CompositeNavigationProp } from '@react-navigation/native';
 import { BottomTabNavigationProp } from '@react-navigation/bottom-tabs';
-import { BottomTabsList } from '../navigation/BottomTabsList';
+import { BottomTabsList } from '../../navigation/BottomTabsList';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import { RootStackParamList } from '../navigation/RootStackedList';
-import { useTheme } from '../../context/ThemeContext';
+import { RootStackParamList } from '../../navigation/RootStackedList';
+import { useTheme } from '../../../context/ThemeContext';
 import { GoogleSignin } from '@react-native-google-signin/google-signin';
 import AsyncStorage from '@react-native-async-storage/async-storage';
